geom_isobands {ggisoband}	R Documentation

Draw isoband and isoline contours

Description

Draw isoband and isoline contours.

Usage

geom_isobands(mapping = NULL, data = NULL, stat = "isolevels",
  position = "identity", ..., bins = NULL, binwidth = NULL,
  breaks = NULL, polygon_outline = TRUE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

geom_density_bands(mapping = NULL, data = NULL, stat = "densitygrid",
  position = "identity", ..., bins = NULL, binwidth = NULL,
  breaks = NULL, polygon_outline = TRUE, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data.

stat
The statistical transformation to use on the data for this layer, as a string.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

bins
Number of bins for discretization. Has priority over binwidth.

binwidth
Binwidth used during discretization.

breaks
Explicit bin boundaries to use for discretization. Has priority over both bins and binwidth.

polygon_outline
Draw filled polygons with equally colored outlines? The default is TRUE, which works well in cases where isobands are drawn without colored isolines. However, it can create drawing artifacts when used in combination with alpha transparency.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

Examples

library(ggplot2)

volcano3d <- reshape2::melt(volcano)
names(volcano3d) <- c("x", "y", "z")

ggplot(volcano3d, aes(x, y, z = z)) +
  geom_isobands(aes(color = stat(zmin)), fill = NA) +
  scale_color_viridis_c() +
  coord_cartesian(expand = FALSE) +
  theme_bw()

ggplot(volcano3d, aes(x, y, z = z)) +
  geom_isobands(aes(fill = stat(zmin)), color = NA) +
  scale_fill_viridis_c(guide = "legend") +
  coord_cartesian(expand = FALSE) +
  theme_bw()

# set polygon_outline = FALSE when drawing filled polygons
# with alpha transparency
ggplot(volcano3d, aes(x, y, z = z)) +
  geom_isobands(
    aes(fill = stat(zmin)), color = NA,
    alpha = 0.5, polygon_outline = FALSE
  ) +
  scale_fill_viridis_c(guide = "legend") +
  coord_cartesian(expand = FALSE) +
  theme_bw()

ggplot(faithful, aes(eruptions, waiting)) +
  geom_density_bands(
    aes(fill = stat(density)),
    color = "gray40", alpha = 0.7, size = 0.2
   ) +
  geom_point(shape = 21, fill = "white") +
  scale_fill_viridis_c(guide = "legend")

[Package ggisoband version 0.0.0.9000 Index]