stat_density_grid {ggisoband}	R Documentation

Generate 2d kernel density estimates

Description

Generate 2d kernel density estimates

Usage

stat_density_grid(mapping = NULL, data = NULL, geom = "tile",
  position = "identity", ..., n = 100, h = NULL, bins = NULL,
  binwidth = NULL, breaks = NULL, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data.

geom
The geometric object to use display the data

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

n
Number of grid points in each direction.

h
Bandwidth (vector of length two). If NULL, estimated using MASS::bandwidth.nrd().

bins
Number of bins for discretization. Has priority over binwidth.

binwidth
Binwidth used during discretization.

breaks
Explicit bin boundaries to use for discretization. Has priority over both bins and binwidth.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

Computed variables

Same as stat_isolevels(), with the addition of:

density

the density estimate

ndensity

density estimate scaled to maximum of 1

z

the density estimate, identical to density

Examples

library(ggplot2)

# default uses `geom_tile()` for drawing
ggplot(faithful, aes(eruptions, waiting)) +
  stat_density_grid(aes(fill = stat(density))) +
  geom_point(size = 0.3, color = "white")

# discretized colors
ggplot(faithful, aes(eruptions, waiting)) +
  stat_density_grid(aes(fill = stat(zmin))) +
  geom_point(size = 0.3, color = "white")

# points
ggplot(faithful, aes(eruptions, waiting)) +
  stat_density_grid(
    aes(color = stat(density), size = stat(density)),
    geom = "point", n = 20, stroke = 0
  ) +
  geom_point(size = 0.75, shape = 21, fill = "white", color = "black")

# contour bands
ggplot(faithful, aes(eruptions, waiting)) +
  stat_density_grid(
    aes(fill = stat(zmin)), geom = "isobands",
    color = "gray70"
  ) +
  geom_point(size = 0.3, color = "white")

[Package ggisoband version 0.0.0.9000 Index]