stat_isolevels {ggisoband}	R Documentation

Bin a 2d elevation raster into bands of similar height

Description

Bin a 2d elevation raster into bands of similar height.

Usage

stat_isolevels(mapping = NULL, data = NULL, geom = "contour",
  position = "identity", ..., bins = NULL, binwidth = NULL,
  breaks = NULL, na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data.

geom
The geometric object to use display the data

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

bins
Number of bins for discretization. Has priority over binwidth.

binwidth
Binwidth used during discretization.

breaks
Explicit bin boundaries to use for discretization. Has priority over both bins and binwidth.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

Computed variables

level

height of contour band, expressed as integer

zmin

minimum z value of contour band

zmax

maximum z value of contour band

nzmax

maximum z value of contour band, normalized to a maximum of 1

[Package ggisoband version 0.0.0.9000 Index]